refactor: input image 컴포넌트화
