+++ src/components/auth/signup/signupForm.tsx
@@ -1,9 +1,8 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
-import { ChangeEvent, useState } from "react";
+import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { Avatar, AvatarFallback, AvatarImage } from "../../ui/avatar";
 import { Button } from "../../ui/button";
 import {
   Form,
@@ -14,6 +13,7 @@ import {
   FormMessage,
 } from "../../ui/form";
 import { Input } from "../../ui/input";
+import InputImage from "./inputImage";
 import { signupFormSchema } from "./signupFormSchema";
 
 const SignupForm = () => {
@@ -29,21 +29,6 @@ const SignupForm = () => {
     },
   });
 
-  function getImageData(event: ChangeEvent<HTMLInputElement>) {
-    // FileList is immutable, so we need to create a new one
-    const dataTransfer = new DataTransfer();
-
-    // Add newly uploaded images
-    Array.from(event.target.files!).forEach((image) =>
-      dataTransfer.items.add(image)
-    );
-
-    const files = dataTransfer.files;
-    const displayUrl = URL.createObjectURL(event.target.files![0]);
-
-    return { files, displayUrl };
-  }
-
   function onSubmit(values: z.infer<typeof signupFormSchema>) {
     console.log(values);
     router.push("/");
@@ -56,34 +41,10 @@ const SignupForm = () => {
           onSubmit={formSignup.handleSubmit(onSubmit)}
           className="w-full flex flex-col gap-6"
         >
-          <FormField
+          <InputImage
             control={formSignup.control}
-            name="image"
-            render={({ field: { onChange, value, ...rest } }) => (
-              <FormItem className="mt-10 mb-8 flex-center flex-col">
-                <FormLabel className="flex-center flex-col gap-4 cursor-pointer">
-                  <Avatar className="size-32">
-                    <AvatarImage src={preview} />
-                    <AvatarFallback>기본</AvatarFallback>
-                  </Avatar>
-                  <p className="text-gray-400 text-xl">사진 추가</p>
-                </FormLabel>
-                <FormControl>
-                  <Input
-                    type="file"
-                    accept="image/*"
-                    className="hidden"
-                    onChange={(event) => {
-                      const { files, displayUrl } = getImageData(event);
-                      setPreview(displayUrl);
-                      onChange(files[0]);
-                    }}
-                    {...rest}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
+            preview={preview}
+            setPreview={setPreview}
           />
           <FormField
             control={formSignup.control}
