+++ src/components/auth/signup/signupFormSchema.ts
@@ -14,5 +14,5 @@ export const signupFormSchema = z.object({
       (file) => file?.size <= IMAGE_SIZE * 1000 * 1000,
       `이미지 최대 사이즈는 ${IMAGE_SIZE}MB 입니다.`
     )
-    .refine((file) => file?.type.includes("image/")),
+    .refine((file: File) => file?.type.includes("image/")),
 });
